major SDK overhaul to support connection to multiple devices
